make ProductImage.create match Magento API
Previously this was a layering violation. There's no reason for `create` to change the argument signature of just this method, base64 encode the image, and detect its mime-type, especially if `update` does not do any of that. The calling code should be responsible for encoding and mime-type detecting in both cases, and both methods should match the Magento API.